sss/eff: use Void i/o () to denote no effects
() is not the same as Void, in that () *does* have a constructor. '()'
has a single constructor and as such 'Eff ()' only means that no effects
are left to handle by convention. 'Void' on the other hand has no
constructors and as such because there really is no way to create a
value of type Void, 'Eff Void' indeed does denote the scenario where no
effects are left to be handled.