develop: fix warnings in src and test
add test for multiple effects (of a similar nature) in a single
computation.

also enable a few extensions (such as MonoLocalBinds) which make type
inference more predictable. in some cases this requires a few extra
annotations in the library functions, but it's fairly straightforward
and doesn't impose much of an overhead during usage.