eff: use GADTs for effect definition
this simplifies the definition of the particular effect datatype,
however, transfers some of the responsibility to the interpreter. case
in point, when we represent 'Writer' effect in this manner, the value
'v' (of corresponding 'Eff r v') is no longer accessible in the
constructor. as such, the interpreter 'runWriter' is responsible both
for deciding the value to be passed to the continuation 'k' ('()' in
this case) as well as passing it (previously the interpreter was only
responsible for passing it, not deciding what it ought to be).