eff/streamline: generalize handle/relay
The Relay typeclass is inductive and abstracts the pattern of relaying
requests. The Handle typeclass defines the 'handle' method. This
'handle' method along with a way to embed pure values (i.e., "return"
which acts as a domain transformer) allows us to interpret effects
in (Eff r).

The "return" method (:: a -> k) is under specified (in terms of types)
and has not been encapsulated in a typeclass. It's unclear if there's
a better way to package 'handle' and 'return' together.

Given these changes, at times GHC needs some hand-holding and we need
to make use of the TypeApplications language extension. As such, we'll
be dropping support for ghc < 8.0 shortly.