eff/logic: port tests from Eff.Logic.Choose
eff/logic: use foldl' variant of msum

eff/logic: reverting to using foldr everywhere

The use of foldl' in choose can result in quadratic times when used with
sols. Using foldr in choose and foldl' in makeChoiceLst seems to be the best. I
believe this is because it allows for some fusion. In short, from best to worse:

- foldr in choose, foldl' in makeChoiceLst: linear + some deforestation
- foldr in choose, foldr in makeChoiceLst: linear
- foldl' in choose, foldl' in makeChoiceLst: quadratic